Add files via upload
added a summary table that draws from a newly created SummarTable class to the TLM tool that creates a separate tab from the report that summarizes each assays calls as Trait, seg, wildtype, no call or fail and the % data return.  Function of the TLM tool on its own is unchanged.